requirements.txt
torch
diffusers
transformers
numpy
matplotlib
Pillow
scipy